+++ internal/terminalui/terminal_other.go
@@ -7,3 +7,7 @@ import "os"
 func enableVirtualTerminal(_ *os.File) bool {
 	return true
 }
+
+func terminalWidth(_ *os.File) int {
+	return 0
+}