+++ internal/terminalui/terminal_windows.go
@@ -16,3 +16,11 @@ func enableVirtualTerminal(output *os.File) bool {
 	}
 	return windows.SetConsoleMode(handle, mode|windows.ENABLE_VIRTUAL_TERMINAL_PROCESSING) == nil
 }
+
+func terminalWidth(output *os.File) int {
+	var info windows.ConsoleScreenBufferInfo
+	if windows.GetConsoleScreenBufferInfo(windows.Handle(output.Fd()), &info) != nil {
+		return 0
+	}
+	return int(info.Window.Right-info.Window.Left) + 1
+}